User may now delete objects using the Del shortcut. But in this case (not when using the menu) a dialogbox is displayed when more than an object is selected

+++ src/core/toolmap.cpp
@@ -117,6 +117,8 @@ BEGIN_EVENT_TABLE (ToolMapFrame, wxFrame)
 	EVT_MENU (ID_MENU_CUT_LINES, ToolMapFrame::OnToolChanged)
 
 	//EDIT MENU
+	EVT_MENU_HIGHLIGHT(ID_MENU_DELETE_OBJ, ToolMapFrame::OnMenuHighlightOn)
+	EVT_MENU_HIGHLIGHT_ALL( ToolMapFrame::OnMenuHighligntOff)
 	EVT_MENU (ID_MENU_DELETE_OBJ, ToolMapFrame::OnEditDeleteSelected)
 	EVT_MENU (ID_MENU_UNDO, ToolMapFrame::OnEditUndo)
 	EVT_MENU (ID_MENU_CREATE_INTERSECTIONS, ToolMapFrame::OnCreateIntersections)
@@ -263,6 +265,7 @@ void ToolMapFrame::PostInit()
 	wxLogMessage(_("wxWidgets version is : %s"), wxVERSION_STRING);
 	wxLogMessage(_("Running under : %s"), wxGetOsDescription().c_str());
 	
+	m_MenuForDeletePressed = false;
 	
 	// loading GIS drivers
 	tmGISData::InitGISDrivers(TRUE, TRUE);
@@ -374,7 +377,7 @@ wxMenuBar* ToolMapFrame::CreateToolMapMenu()
     itemMenu41->Append(ID_MENU_MERGE_LINES, _("Merge selected lines\tCtrl+M"), _T(""), wxITEM_NORMAL);
     itemMenu41->Append(ID_MENU_CREATE_INTERSECTIONS, _("Create intersections\tCtrl+I"), _T(""), wxITEM_NORMAL);
     itemMenu41->Append(ID_MENU_FLIP_LINE, _("Flip line\tCtrl+F"), _T(""), wxITEM_NORMAL);
-    itemMenu41->Append(ID_MENU_DELETE_OBJ, _("Delete selected objects"), _T(""), wxITEM_NORMAL);
+    itemMenu41->Append(ID_MENU_DELETE_OBJ, _("Delete selected objects\tDel"), _T(""), wxITEM_NORMAL);
     itemMenu41->Append(ID_MENU_ADJUST_SNAPPING, _("Adjust snapping...\tCtrl+S"), _T(""), wxITEM_CHECK);
     menuBar->Append(itemMenu41, _("Edition"));
     wxMenu* itemMenu55 = new wxMenu;
@@ -790,6 +793,35 @@ void ToolMapFrame::OnEditSwitch (wxCommandEvent & event)
  *******************************************************************************/
 void ToolMapFrame::OnEditDeleteSelected (wxCommandEvent & event)
 {
+	/*wxString myStatus = _T("false");
+	if (m_MenuForDeletePressed == true)
+		myStatus = _T("true");
+	
+		wxLogDebug(_T("Menu is pressed : ") + myStatus);*/
+	
+	bool myMenuPressed = m_MenuForDeletePressed;
+	m_MenuForDeletePressed = false;
+
+	
+	if (m_EditManager->IsDrawingAllowed()==false)
+	{
+		return;
+	}
+	
+	// display dialog if user use shortcut	
+	if(myMenuPressed == false)
+	{
+		int iSelNum = m_EditManager->GetSelectionCount();
+		if (iSelNum > 1)
+			
+			if (wxMessageBox(wxString::Format(_("%d Objects selected ! Confirm deleting ?"), iSelNum),
+							 wxString::Format(_("Deleting %d objects"), iSelNum),
+							 wxICON_WARNING | wxYES_NO, this) ==wxNO)
+				return;
+		
+	}
+	
+	
 	m_EditManager->DeleteSelected(true);
 }
 
@@ -865,6 +897,32 @@ void ToolMapFrame::OnShowInformationDialog (wxCommandEvent & event)
 	*pi = 12;*/
 	
 	m_AttribManager->DisplayInformationsWnd();
+}
+
+
+/***************************************************************************//**
+ @brief Called when menu Delete is highlited
+ @author Lucien Schreiber (c) CREALP 2009
+ @date 12 May 2009
+ *******************************************************************************/
+void ToolMapFrame::OnMenuHighlightOn (wxMenuEvent & event)
+{
+	m_MenuForDeletePressed = true;
+	event.Skip();
+}
+
+
+/***************************************************************************//**
+ @brief Called when menu are closed
+ @author Lucien Schreiber (c) CREALP 2009
+ @date 12 May 2009
+ *******************************************************************************/
+void ToolMapFrame::OnMenuHighligntOff(wxMenuEvent & event)
+{
+
+	if (m_MenuForDeletePressed == true && event.GetMenuId() != ID_MENU_DELETE_OBJ)
+		m_MenuForDeletePressed = false;
+	event.Skip();
 }
